同步 ead0163f4e42079e08e018f89b7d4b1caa7a5b27  performance optimization  mm及bmm适配代码优化

+++ src/aten/src/ATen/native/npu/CumsumKernelNpu.cpp
@@ -25,7 +25,7 @@ Tensor& _cumsum_out_npu(Tensor& result, const Tensor& self, int64_t dim) {
       Scalar dimScalar(dim);
       cmd.Name("Cumsum")
       .Input(self)
-      .Input(dimScalar, at::kInt)
+      .Input(dimScalar, at::kLong)
       .Output(result)
       .Run();
     } else {
